Time: 54 ms (34.95%), Space: 17.9 MB (66.04%) - LeetHub

+++ 0268-missing-number/0268-missing-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int missingNumber(vector<int>& nums) {
-        long long res = nums.size() * (nums.size() + 1) / 2;
+        int res = nums.size() * (nums.size() + 1) / 2;
         for (auto x : nums)
             res -= x;
         return res;
